Add Blog list component
