+++ src/pages/Blog.jsx
@@ -1,10 +1,12 @@
 import React from "react";
 import TitleBgImage from "../components/UI/TitleBgImage";
+import BlogList from "../components/Blog/BlogList";
 
 const Blog = () => {
   return (
     <>
       <TitleBgImage>Blog</TitleBgImage>
+      <BlogList />
     </>
   );
 };
